Task 2: Corrections

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -1,15 +1,15 @@
 #include "main.h"
 
 /**
- * _strlen - Returns the length of a string.
- * @str: nput char.
+ * _strlen - Length of a string
+ * @str: Input char.
  *
- * Return: The length of a string.
+ * Return: The length of @str
  */
 
 int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*s++)
 		length++;
